huggingface/requirements.txt
gradio==4.44.0
ultralytics==8.3.225
pillow==10.1.0
opencv-python-headless==4.10.0.84
torch>=2.1.0
torchvision>=0.16.0
